Remove Nine Patch layers.
These were only used for doing effects in Clank and the
shadows around the page when bouncing on the Mac. Now
they're dead code.

In the future if we decide nine patch is an API we really
want we should expose web platform primitives instead of
having a special API on layer.

R=ojan@chromium.org, abarth@chromium.org

Review URL: https://codereview.chromium.org/685983002