Only pointers which are down should be tracked.
This change ensures that we only store the results of a hit
test on the initial pointer down event.  Moreover, we perform
new hit tests each time a hovering pointer moves.  This is
important to ensure correct behavior of input devices which can
hover, such as mice.

Previously the first hover movement after releasing a mouse
button would cause a new pointer state to be recorded along with
hit test results for wherever the pointer happened to be which
caused the following pointer down event to be delivered to the
wrong place.

Fixes issue #1189.