Allow creation of symlinks for Goma via an environment variable. (#23765)
Setting `FLUTTER_GOMA_CREATE_XCODE_SYMLINKS=1` as an environment variable should
achieve the same result as manually specifying the `--xcode-symlinks` flag
(which I often forget). This is in keeping with the pattern of setting the
`GOMA_DIR` environment variable. The variable has a the `FLUTTER_` prefix
because environment variable with the `GOMA_` prefix are reserved for use by
Goma and variable unknown to Goma will cause a fatal error.