Use flutter mirrors for non-google origin deps instead of fuchsia (#48735)
I have [configured mirrors](https://flutter.googlesource.com/third_party) with the same configuration as other flutter mirrors (e.g. public read access, hosted on flutter GoB) for the fuchsia github dependencies coming from non-google origin. The same patches are applied to these as in fuchsia GoB.

This change is to rely on the flutter mirrors instead of fuchsia so that possible breaking changes made to fuchsia mirrors in the future will not affect flutter. Context and tracking can be found in this bug: b/313921121 

Github deps from google origins are not being migrated to a new flutter mirror here because google is creating its own mirrors for those deps. We will migrate deps included in the new mirrors after it is set up. (context: b/274009025)

[C++, Objective-C, Java style guides]: https://github.com/flutter/engine/blob/main/CONTRIBUTING.md#style